feat(형진) : 기수 테이블 완료

See merge request s09-webmobile4-sub2/S09P12D103!26

+++ backend/ASAF/src/main/java/com/ASAF/dto/GenerationDTO.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class GenerationDTO {
-    private Long generation_code;
+    private int generation_code;
     private String generation_num;
 
     public static GenerationDTO toGenerationDTO(GenerationEntity generationEntity) {
